Move applications utilities in a separate library
