+++ src/CMakeLists.txt
@@ -2,4 +2,5 @@ add_subdirectory(afes)
 
 if(BUILD_APPS)
 add_subdirectory(apps)
+add_subdirectory(apps_utils)
 endif()
