+++ src/apps/AdaptiveTreeVisualization.cpp
@@ -33,13 +33,13 @@ DAMAGE.
 #include <math.h>
 #include <float.h>
 #include <afes/MyMiscellany.h>
-#include <afes/CmdLineParser.h>
+#include <apps_utils/CmdLineParser.h>
+#include <apps_utils/io/Image.h>
 #include <afes/PPolynomial.h>
 #include <afes/FEMTree.h>
 #include <afes/Ply.h>
 #include <afes/VertexFactory.h>
 #include <afes/RegularGrid.h>
-#include <afes/Image.h>
 
 cmdLineParameter< char* >
 	In( "in" ) ,
