+++ src/apps/ChunkPLY.cpp
@@ -37,8 +37,8 @@ DAMAGE.
 #include <limits>
 #include <sstream>
 #include <unordered_map>
+#include <apps_utils/CmdLineParser.h>
 #include <afes/MyMiscellany.h>
-#include <afes/CmdLineParser.h>
 #include <afes/Geometry.h>
 #include <afes/Ply.h>
 #include <afes/VertexStream.h>
