+++ src/apps/EDTInHeat.cpp
@@ -37,8 +37,8 @@ DAMAGE.
 #include <math.h>
 #include <float.h>
 #include <functional>
+#include <apps_utils/CmdLineParser.h>
 #include <afes/MyMiscellany.h>
-#include <afes/CmdLineParser.h>
 #include <afes/PPolynomial.h>
 #include <afes/FEMTree.h>
 #include <afes/Ply.h>
