+++ src/apps/SurfaceTrimmer.cpp
@@ -34,9 +34,9 @@ DAMAGE.
 #include <stdlib.h>
 #include <float.h>
 #include <algorithm>
+#include <apps_utils/CmdLineParser.h>
 #include <afes/FEMTree.h>
 #include <afes/MyMiscellany.h>
-#include <afes/CmdLineParser.h>
 #include <afes/MAT.h>
 #include <afes/Geometry.h>
 #include <afes/Ply.h>
