+++ src/apps_utils/include/apps_utils/io/Image.h
@@ -4,7 +4,7 @@
 #define SUPPORT_TILES
 
 #include <string.h>
-#include "MyMiscellany.h"
+#include <afes/MyMiscellany.h>
 
 struct ImageReader
 {
